Do not allow user to perform actions if he is not authorized

+++ src/app/words/actions.ts
@@ -1,6 +1,7 @@
 "use server";
 import { Word } from "@prisma/client";
 import { revalidatePath } from "next/cache";
+import { auth } from "@clerk/nextjs/server";
 
 import { prisma } from "@/core/prisma";
 import { itemsPerPage } from "@/core/constants";
@@ -14,6 +15,7 @@ type Metrics = {
 };
 
 const getErrorMessage = (submessage: string) => `There was an error while ${submessage}`;
+const unauthorisedError = "You are not authorised to ";
 
 // Add a new word
 async function addEditWord(formData: FormData, selectedClass = "", wordToEditId = ""): Promise<WordResponse> {
@@ -25,6 +27,10 @@ async function addEditWord(formData: FormData, selectedClass = "", wordToEditId 
 
   if (!spanish || !english || !greek || !selectedClass)
     return { error: "Spanish english, greek words and class are required" };
+
+  // Get logged in user
+  const { userId } = await auth();
+  if (!userId) return { error: `${unauthorisedError} add or edit word` };
 
   try {
     if (wordToEditId) {
@@ -71,6 +77,9 @@ async function getWords(page: number): Promise<{ words?: Word[]; error?: string 
 
 // Mark a word as learned
 async function markAsLearned(word: Word): Promise<WordResponse> {
+  const { userId } = await auth();
+  if (!userId) return { error: `${unauthorisedError} mark word as learned` };
+
   const newLearnedValue = !word.learned;
   try {
     await prisma.word.update({
@@ -86,6 +95,9 @@ async function markAsLearned(word: Word): Promise<WordResponse> {
 }
 
 async function highlighWord(word: Word): Promise<WordResponse> {
+  const { userId } = await auth();
+  if (!userId) return { error: `${unauthorisedError} highlight word` };
+
   const newHighlightValue = !word.highlight;
   try {
     await prisma.word.update({
@@ -102,6 +114,9 @@ async function highlighWord(word: Word): Promise<WordResponse> {
 
 // Delete a word
 async function deleteWord(id: string): Promise<WordResponse> {
+  const { userId } = await auth();
+  if (!userId) return { error: `${unauthorisedError} delete word` };
+
   try {
     await prisma.word.delete({ where: { id } });
     revalidatePath("/");
